service response cleanup.

+++ src/main/java/com/javee/attendance/controllers/employee/EmployeeController.java
@@ -1,12 +1,11 @@
 package com.javee.attendance.controllers.employee;
 
+import com.javee.attendance.controllers.BaseController;
 import com.javee.attendance.entities.Employee;
-import com.javee.attendance.entities.User;
 import com.javee.attendance.model.JWTUserDetails;
 import com.javee.attendance.repositories.EmployeeRepository;
 import io.swagger.annotations.Api;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
@@ -16,7 +15,7 @@ import java.util.Optional;
 @RestController
 @Api( value = "Employee", description = "REST API for Employee", tags = { "Employee" } )
 @RequestMapping( "/api" )
-public class EmployeeController
+public class EmployeeController extends BaseController
 {
 	@Autowired
 	private EmployeeRepository employeeRepository;
@@ -24,20 +23,22 @@ public class EmployeeController
 	@CrossOrigin
 	@RequestMapping( value = "/employee", method = RequestMethod.POST,
 			produces = "application/json", consumes = "application/json" )
-	public Employee createEmpolyee( @AuthenticationPrincipal JWTUserDetails jwtUserDetails, @RequestBody Employee employee )
+	public ResponseEntity<Employee> createEmpolyee( @AuthenticationPrincipal JWTUserDetails jwtUserDetails, @RequestBody Employee employee )
 	{
-		System.out.println( "User Id : " + jwtUserDetails.getId() );
 		employee = employeeRepository.save( employee );
-		return employee;
+		return generateOkResponse( employee );
 	}
 
 	@CrossOrigin
 	@RequestMapping( value = "/employee/{id}", method = RequestMethod.GET,
 			produces = "application/json" )
-	public Employee getEmployeeById( @AuthenticationPrincipal JWTUserDetails jwtUserDetails, @PathVariable( "id" ) Long id )
+	public ResponseEntity<Employee> getEmployeeById( @AuthenticationPrincipal JWTUserDetails jwtUserDetails, @PathVariable( "id" ) Long id )
 	{
+		if (id == null || id == 0)
+			return generateBadRequestResponse();
+
 		Optional<Employee> employeeOptional = employeeRepository.findById( id );
-		return employeeOptional.get();
+		return generateOkResponse( employeeOptional.get() );
 	}
 
 	@CrossOrigin
@@ -45,12 +46,7 @@ public class EmployeeController
 			produces = "application/json" )
 	public ResponseEntity<List<Employee>> getEmployees( @AuthenticationPrincipal JWTUserDetails jwtUserDetails )
 	{
-		System.out.println( "User Id : " + jwtUserDetails.getId() + " User Name : " + jwtUserDetails.getUsername());
-		if(!jwtUserDetails.getAuthorities().stream().findFirst().get().toString().equalsIgnoreCase( User.ROLE.ADMIN.name()) ){
-			return new ResponseEntity<>(  HttpStatus.FORBIDDEN );
-		}
-
-		return new ResponseEntity<>( employeeRepository.findAll(), HttpStatus.OK );
+		return generateOkResponse( employeeRepository.findAll() );
 	}
 
 	@CrossOrigin
@@ -58,21 +54,27 @@ public class EmployeeController
 			produces = "application/json", consumes = "application/json" )
 	public ResponseEntity<Object> updateEmployee( @AuthenticationPrincipal JWTUserDetails jwtUserDetails, @RequestBody Employee employee, @PathVariable Long id )
 	{
+		if (id == null || id == 0)
+			return generateBadRequestResponse();
+
 		Optional<Employee> employeeOptional = employeeRepository.findById( id );
 		if (!employeeOptional.isPresent())
-			return ResponseEntity.notFound().build();
+			return generateNotFoundResponse();
 		employee.setId( id );
 		employeeRepository.save( employee );
-		return ResponseEntity.noContent().build();
+		return generateNoContentResponse();
 	}
 
 	@CrossOrigin
 	@RequestMapping( value = "/employee/{id}", method = RequestMethod.DELETE,
 			produces = "application/json" )
-	public boolean deleteEmployeeById( @AuthenticationPrincipal JWTUserDetails jwtUserDetails, @PathVariable( "id" ) Long id )
+	public ResponseEntity<Object> deleteEmployeeById( @AuthenticationPrincipal JWTUserDetails jwtUserDetails, @PathVariable( "id" ) Long id )
 	{
+		if (id == null || id == 0)
+			return generateBadRequestResponse();
+
 		employeeRepository.deleteById( id );
-		return true;
+		return generateNoContentResponse();
 
 	}
