+++ src/main/java/com/javee/attendance/controllers/user/UserController.java
@@ -1,5 +1,6 @@
 package com.javee.attendance.controllers.user;
 
+import com.javee.attendance.controllers.BaseController;
 import com.javee.attendance.entities.User;
 import com.javee.attendance.model.JWTUserDetails;
 import com.javee.attendance.repositories.UserRepository;
@@ -18,7 +19,7 @@ import java.util.Optional;
 @RestController
 @Api( value = "User", description = "Endpoint for User", tags = { "User" } )
 @RequestMapping( "/api" )
-public class UserController
+public class UserController extends BaseController
 {
 	private static final Logger LOGGER = LoggerFactory.getLogger( UserController.class );
 
@@ -28,31 +29,34 @@ public class UserController
 	@CrossOrigin
 	@RequestMapping( value = "/user", method = RequestMethod.POST,
 			produces = "application/json", consumes = "application/json" )
-	public User createUser( @AuthenticationPrincipal JWTUserDetails jwtUserDetails, @RequestBody User user )
+	public ResponseEntity<User> createUser( @AuthenticationPrincipal JWTUserDetails jwtUserDetails, @RequestBody User user )
 	{
 		String password = user.getPassword();
 		user.setPassword( PasswordEncryptor.encrypt( password ) );
 		user = userRepository.save( user );
-		return user;
+		return generateOkResponse( user );
 	}
 
 	@CrossOrigin
 	@ApiOperation( value = "Get User", tags = { "User" } )
 	@RequestMapping( value = "/user/{id}", method = RequestMethod.GET,
 			produces = "application/json" )
-	public User getUserById( @AuthenticationPrincipal JWTUserDetails jwtUserDetails, @PathVariable( "id" ) Long id )
+	public ResponseEntity<User> getUserById( @AuthenticationPrincipal JWTUserDetails jwtUserDetails, @PathVariable( "id" ) Long id )
 	{
+		if (id == null || id == 0)
+			return generateBadRequestResponse();
+
 		Optional<User> userOptional = userRepository.findById( id );
-		return userOptional.get();
+		return generateOkResponse( userOptional.get() );
 	}
 
 	@CrossOrigin
 	@ApiOperation( value = "Get List of Users", tags = { "User" } )
 	@RequestMapping( value = "/users", method = RequestMethod.GET,
 			produces = "application/json" )
-	public List<User> getUsers( @AuthenticationPrincipal JWTUserDetails jwtUserDetails )
+	public ResponseEntity<List<User>> getUsers( @AuthenticationPrincipal JWTUserDetails jwtUserDetails )
 	{
-		return userRepository.findAll();
+		return generateOkResponse( userRepository.findAll() );
 	}
 
 	@CrossOrigin
@@ -61,25 +65,30 @@ public class UserController
 			produces = "application/json", consumes = "application/json" )
 	public ResponseEntity<Object> updateUser( @AuthenticationPrincipal JWTUserDetails jwtUserDetails, @RequestBody User user, @PathVariable Long id )
 	{
+		if (id == null || id == 0)
+			return generateBadRequestResponse();
+
 		Optional<User> userOptional = userRepository.findById( id );
 		if (!userOptional.isPresent())
-			return ResponseEntity.notFound().build();
+			return generateNotFoundResponse();
 		user.setId( id );
 		String password = user.getPassword();
 		user.setPassword( PasswordEncryptor.encrypt( password ) );
 		userRepository.save( user );
-		return ResponseEntity.noContent().build();
+		return generateNoContentResponse();
 	}
 
 	@CrossOrigin
 	@ApiOperation( value = "Delete User", tags = { "User" } )
 	@RequestMapping( value = "/user/{id}", method = RequestMethod.DELETE,
 			produces = "application/json" )
-	public boolean deleteUserById( @AuthenticationPrincipal JWTUserDetails jwtUserDetails, @PathVariable( "id" ) Long id )
+	public ResponseEntity<Object> deleteUserById( @AuthenticationPrincipal JWTUserDetails jwtUserDetails, @PathVariable( "id" ) Long id )
 	{
-		userRepository.deleteById( id );
-		return true;
+		if (id == null || id == 0)
+			return generateBadRequestResponse();
 
+		userRepository.deleteById( id );
+		return generateNoContentResponse();
 	}
 }
